requirements.txt
Flask==2.3.2
google-cloud-firestore==2.11.0
gunicorn==20.1.0
Jinja2==3.1.2
Werkzeug==2.3.7
